Move statedb initialization outside
