+++ src/Nethermind/Nethermind.Init/Steps/InitializeBlockchain.cs
@@ -1,18 +1,13 @@
 // SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-using System;
 using System.Collections.Generic;
-using System.IO.Abstractions;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Nethermind.Api;
 using Nethermind.Blockchain;
 using Nethermind.Blockchain.Filters;
-using Nethermind.Blockchain.FullPruning;
 using Nethermind.Blockchain.Services;
-using Nethermind.Blockchain.Synchronization;
 using Nethermind.Config;
 using Nethermind.Consensus;
 using Nethermind.Consensus.Comparers;
@@ -21,30 +16,16 @@ using Nethermind.Consensus.Producers;
 using Nethermind.Consensus.Validators;
 using Nethermind.Core;
 using Nethermind.Core.Attributes;
-using Nethermind.Core.Crypto;
-using Nethermind.Core.Extensions;
-using Nethermind.Db;
-using Nethermind.Db.FullPruning;
 using Nethermind.Evm;
 using Nethermind.Evm.TransactionProcessing;
-using Nethermind.JsonRpc.Converters;
-using Nethermind.JsonRpc.Modules.DebugModule;
 using Nethermind.JsonRpc.Modules.Eth.GasPrice;
-using Nethermind.JsonRpc.Modules.Trace;
 using Nethermind.Logging;
-using Nethermind.Serialization.Json;
-using Nethermind.State;
-using Nethermind.State.Witnesses;
-using Nethermind.Synchronization.Trie;
-using Nethermind.Synchronization.Witness;
-using Nethermind.Trie;
-using Nethermind.Trie.Pruning;
 using Nethermind.TxPool;
 using Nethermind.Wallet;
 
 namespace Nethermind.Init.Steps
 {
-    [RunnerStepDependencies(typeof(InitializePlugins), typeof(InitializeBlockTree), typeof(InitializeContainer), typeof(SetupKeyStore))]
+    [RunnerStepDependencies(typeof(InitializePlugins), typeof(InitializeBlockTree), typeof(InitializeContainer), typeof(SetupKeyStore), typeof(InitializeStateDb))]
     public class InitializeBlockchain : IStep
     {
         private readonly INethermindApi _api;
@@ -64,8 +45,6 @@ namespace Nethermind.Init.Steps
         [Todo(Improve.Refactor, "Use chain spec for all chain configuration")]
         protected virtual Task InitBlockchain()
         {
-            InitBlockTraceDumper();
-
             (IApiWithStores getApi, IApiWithBlockchain setApi) = _api.ForBlockchain;
 
             if (getApi.ChainSpec is null) throw new StepDependencyException(nameof(getApi.ChainSpec));
@@ -75,133 +54,8 @@ namespace Nethermind.Init.Steps
 
             _logger = getApi.LogManager.GetClassLogger();
             IInitConfig initConfig = getApi.Config<IInitConfig>();
-            ISyncConfig syncConfig = getApi.Config<ISyncConfig>();
-            IPruningConfig pruningConfig = getApi.Config<IPruningConfig>();
             IBlocksConfig blocksConfig = getApi.Config<IBlocksConfig>();
-            IMiningConfig miningConfig = getApi.Config<IMiningConfig>();
-
-            if (syncConfig.DownloadReceiptsInFastSync && !syncConfig.DownloadBodiesInFastSync)
-            {
-                if (_logger.IsWarn) _logger.Warn($"{nameof(syncConfig.DownloadReceiptsInFastSync)} is selected but {nameof(syncConfig.DownloadBodiesInFastSync)} - enabling bodies to support receipts download.");
-                syncConfig.DownloadBodiesInFastSync = true;
-            }
-
-            IWitnessCollector witnessCollector;
-            if (syncConfig.WitnessProtocolEnabled)
-            {
-                WitnessCollector witnessCollectorImpl = new(getApi.DbProvider.WitnessDb, _api.LogManager);
-                witnessCollector = setApi.WitnessCollector = witnessCollectorImpl;
-                setApi.WitnessRepository = witnessCollectorImpl.WithPruning(getApi.BlockTree!, getApi.LogManager);
-            }
-            else
-            {
-                witnessCollector = setApi.WitnessCollector = NullWitnessCollector.Instance;
-                setApi.WitnessRepository = NullWitnessCollector.Instance;
-            }
-
-            CachingStore cachedStateDb = getApi.DbProvider.StateDb
-                .Cached(Trie.MemoryAllowance.TrieNodeCacheCount);
-            setApi.MainStateDbWithCache = cachedStateDb;
-            IKeyValueStore codeDb = getApi.DbProvider.CodeDb
-                .WitnessedBy(witnessCollector);
-
-            IKeyValueStoreWithBatching stateWitnessedBy = setApi.MainStateDbWithCache.WitnessedBy(witnessCollector);
-            IPersistenceStrategy persistenceStrategy;
-            IPruningStrategy pruningStrategy;
-            if (pruningConfig.Mode.IsMemory())
-            {
-                persistenceStrategy = Persist.IfBlockOlderThan(pruningConfig.PersistenceInterval); // TODO: this should be based on time
-                if (pruningConfig.Mode.IsFull())
-                {
-                    PruningTriggerPersistenceStrategy triggerPersistenceStrategy = new((IFullPruningDb)getApi.DbProvider!.StateDb, getApi.BlockTree!, getApi.LogManager);
-                    getApi.DisposeStack.Push(triggerPersistenceStrategy);
-                    persistenceStrategy = persistenceStrategy.Or(triggerPersistenceStrategy);
-                }
-
-                pruningStrategy = Prune.WhenCacheReaches(pruningConfig.CacheMb.MB()); // TODO: memory hint should define this
-            }
-            else
-            {
-                pruningStrategy = No.Pruning;
-                persistenceStrategy = Persist.EveryBlock;
-            }
-
-            TrieStore trieStore = syncConfig.TrieHealing
-                ? new HealingTrieStore(
-                    stateWitnessedBy,
-                    pruningStrategy,
-                    persistenceStrategy,
-                    getApi.LogManager)
-                : new TrieStore(
-                    stateWitnessedBy,
-                    pruningStrategy,
-                    persistenceStrategy,
-                    getApi.LogManager);
-            setApi.TrieStore = trieStore;
-
-            IWorldState worldState = setApi.WorldState = syncConfig.TrieHealing
-                ? new HealingWorldState(
-                    trieStore,
-                    codeDb,
-                    getApi.LogManager)
-                : new WorldState(
-                    trieStore,
-                    codeDb,
-                    getApi.LogManager);
-
-            if (pruningConfig.Mode.IsFull())
-            {
-                IFullPruningDb fullPruningDb = (IFullPruningDb)getApi.DbProvider!.StateDb;
-                fullPruningDb.PruningStarted += (_, args) =>
-                {
-                    cachedStateDb.PersistCache(args.Context);
-                    trieStore.PersistCache(args.Context, args.Context.CancellationTokenSource.Token);
-                };
-            }
-
-            TrieStoreBoundaryWatcher trieStoreBoundaryWatcher = new(trieStore, _api.BlockTree!, _api.LogManager);
-            getApi.DisposeStack.Push(trieStoreBoundaryWatcher);
-            getApi.DisposeStack.Push(trieStore);
-
-            ITrieStore readOnlyTrieStore = setApi.ReadOnlyTrieStore = trieStore.AsReadOnly(cachedStateDb);
-
-            ReadOnlyDbProvider readOnly = new(getApi.DbProvider, false);
-
-            IStateReader stateReader = setApi.StateReader = new StateReader(readOnlyTrieStore, readOnly.GetDb<IDb>(DbNames.Code), getApi.LogManager);
-
             setApi.TransactionComparerProvider = new TransactionComparerProvider(getApi.SpecProvider!, getApi.BlockTree.AsReadOnly());
-            setApi.ChainHeadStateProvider = new ChainHeadReadOnlyStateProvider(getApi.BlockTree, stateReader);
-
-            worldState.StateRoot = getApi.BlockTree!.Head?.StateRoot ?? Keccak.EmptyTreeHash;
-
-            if (_api.Config<IInitConfig>().DiagnosticMode == DiagnosticMode.VerifyTrie)
-            {
-                Task.Run(() =>
-                {
-                    try
-                    {
-                        _logger!.Info("Collecting trie stats and verifying that no nodes are missing...");
-                        TrieStore noPruningStore = new(stateWitnessedBy, No.Pruning, Persist.EveryBlock, getApi.LogManager);
-                        IWorldState diagStateProvider = new WorldState(noPruningStore, codeDb, getApi.LogManager)
-                        {
-                            StateRoot = getApi.BlockTree!.Head?.StateRoot ?? Keccak.EmptyTreeHash
-                        };
-                        TrieStats stats = diagStateProvider.CollectStats(getApi.DbProvider.CodeDb, _api.LogManager);
-                        _logger.Info($"Starting from {getApi.BlockTree.Head?.Number} {getApi.BlockTree.Head?.StateRoot}{Environment.NewLine}" + stats);
-                    }
-                    catch (Exception ex)
-                    {
-                        _logger!.Error(ex.ToString());
-                    }
-                });
-            }
-
-            // Init state if we need system calls before actual processing starts
-            if (getApi.BlockTree!.Head?.StateRoot is not null)
-            {
-                worldState.StateRoot = getApi.BlockTree.Head.StateRoot;
-            }
-
             setApi.TxValidator = new TxValidator(_api.SpecProvider!.ChainId);
 
             ITxPool txPool = _api.TxPool = CreateTxPool();
@@ -223,7 +77,7 @@ namespace Nethermind.Init.Steps
             setApi.BlockValidator = CreateBlockValidator();
 
             IChainHeadInfoProvider chainHeadInfoProvider =
-                new ChainHeadInfoProvider(getApi.SpecProvider, getApi.BlockTree, stateReader);
+                new ChainHeadInfoProvider(getApi.SpecProvider, getApi.BlockTree, setApi.StateReader!);
 
             // TODO: can take the tx sender from plugin here maybe
             ITxSigner txSigner = new WalletTxSigner(getApi.Wallet, getApi.SpecProvider.ChainId);
@@ -241,7 +95,7 @@ namespace Nethermind.Init.Steps
                 getApi.BlockTree,
                 mainBlockProcessor,
                 _api.BlockPreprocessor,
-                stateReader,
+                setApi.StateReader!,
                 getApi.LogManager,
                 new BlockchainProcessor.Options
                 {
@@ -259,8 +113,6 @@ namespace Nethermind.Init.Steps
             setApi.FilterManager = new FilterManager(filterStore, mainBlockProcessor, txPool, getApi.LogManager);
             setApi.HealthHintService = CreateHealthHintService();
             setApi.BlockProductionPolicy = CreateBlockProductionPolicy();
-
-            InitializeFullPruning(pruningConfig, initConfig, _api, stateReader);
 
             return Task.CompletedTask;
         }
@@ -310,56 +162,6 @@ namespace Nethermind.Init.Steps
                 blockhashProvider,
                 _api.SpecProvider,
                 _api.LogManager);
-        }
-
-        private static void InitializeFullPruning(
-            IPruningConfig pruningConfig,
-            IInitConfig initConfig,
-            INethermindApi api,
-            IStateReader stateReader)
-        {
-            IPruningTrigger? CreateAutomaticTrigger(string dbPath)
-            {
-                long threshold = pruningConfig.FullPruningThresholdMb.MB();
-
-                switch (pruningConfig.FullPruningTrigger)
-                {
-                    case FullPruningTrigger.StateDbSize:
-                        return new PathSizePruningTrigger(dbPath, threshold, api.TimerFactory, api.FileSystem);
-                    case FullPruningTrigger.VolumeFreeSpace:
-                        return new DiskFreeSpacePruningTrigger(dbPath, threshold, api.TimerFactory, api.FileSystem);
-                    default:
-                        return null;
-                }
-            }
-
-            if (pruningConfig.Mode.IsFull())
-            {
-                IDb stateDb = api.DbProvider!.StateDb;
-                if (stateDb is IFullPruningDb fullPruningDb)
-                {
-                    string pruningDbPath = fullPruningDb.GetPath(initConfig.BaseDbPath);
-                    IPruningTrigger? pruningTrigger = CreateAutomaticTrigger(pruningDbPath);
-                    if (pruningTrigger is not null)
-                    {
-                        api.PruningTrigger.Add(pruningTrigger);
-                    }
-
-                    IDriveInfo? drive = api.FileSystem.GetDriveInfos(pruningDbPath).FirstOrDefault();
-                    FullPruner pruner = new(fullPruningDb, api.PruningTrigger, pruningConfig, api.BlockTree!,
-                        stateReader, api.ProcessExit!, ChainSizes.CreateChainSizeInfo(api.ChainSpec.ChainId),
-                        drive, api.LogManager);
-                    api.DisposeStack.Push(pruner);
-                }
-            }
-        }
-
-        private static void InitBlockTraceDumper()
-        {
-            BlockTraceDumper.Converters.AddRange(EthereumJsonSerializer.CommonConverters);
-            BlockTraceDumper.Converters.AddRange(DebugModuleFactory.Converters);
-            BlockTraceDumper.Converters.AddRange(TraceModuleFactory.Converters);
-            BlockTraceDumper.Converters.Add(new TxReceiptConverter());
         }
 
         protected virtual IHealthHintService CreateHealthHintService() =>
